pyproject.toml
[project]
name = "nb-translate-commit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "argparse>=1.4.0",
    "langgraph>=0.0.26",
    "nbformat>=5.7.0",
    "openai>=1.3.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]
